# AnotherMe/packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnotherMe Gateway API",
    "version": "0.3.0",
    "description": "Unified backend gateway for AnotherMe Web and Mobile clients.\n\nCapabilities follow the registry defined in `agents/foundation/capability_registry.py`.\nValid capability ids: `course_generate`, `problem_video_generate`, `quiz_practice`,\n`interactive_demo`, `ai_tutor_chat`, `deep_solve`, `deep_research`,\n`math_animator`, `visualize`, `co_writer`.\n\nP1 changes:\n  - Capability endpoints (P0-3) are now wired.\n  - Global error envelope (`{ error_code, message, request_id, details? }`).\n  - Live-book / co-writer / courses routes are first-class.\n\nAuth: Phase 0 uses static `Bearer <GATEWAY_API_TOKEN>`. Phase 2 will switch to Supabase JWT.",
    "contact": {
      "name": "AnotherMe",
      "url": "https://github.com/anotherme/anotherme"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Phase 0: static token from GATEWAY_API_TOKEN env var. Phase 2: Supabase JWT."
      }
    },
    "schemas": {
...
      "ConversationSummary": {
        "type": "object",
        "required": ["conversation_id", "type", "name", "creator_id", "created_at", "updated_at"],
        "properties": {
          "conversation_id": { "type": "string" },
          "type": { "type": "string" },
          "name": { "type": "string" },
          "creator_id": { "type": "string" },
          "last_message_id": { "type": "string" },
          "last_message_time": { "type": "string" },
          "unread_count": { "type": "integer" },
          "created_at": { "type": "string" },
          "updated_at": { "type": "string" }
        }
      },
      "CreateConversationRequest": {
        "type": "object",
        "required": ["user_id", "name"],
        "properties": {
          "user_id": { "type": "string" },
          "type": { "type": "string", "default": "single" },
          "name": { "type": "string" },
          "creator_id": { "type": "string" },
          "member_ids": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/messages/conversations": {
      "get": {
        "tags": ["messages"],
        "summary": "List conversations for a user",
        "parameters": [
          { "name": "user_id", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "Conversations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ConversationSummary" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["messages"],
        "summary": "Create a conversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateConversationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversation created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ConversationSummary" }
              }
            }
          }
        }
      }
    },
...
  }
}